Rename VesselList component folder
