+++ frontend/src/App.tsx
@@ -28,7 +28,7 @@ import { SideWindow } from './features/SideWindow'
 import { AlertsMapButton } from './features/SideWindow/alerts_reportings/AlertsMapButton'
 import { BeaconMalfunctionsMapButton } from './features/SideWindow/beacon_malfunctions/BeaconMalfunctionsMapButton'
 import { SideWindowLauncher } from './features/SideWindow/SideWindowLauncher'
-import { VesselList } from './features/vessel_list/VesselList'
+import { VesselList } from './features/VesselList'
 import { VesselSidebar } from './features/VesselSidebar'
 import UpdatingVesselLoader from './features/VesselSidebar/UpdatingVesselLoader'
 import { VesselSidebarHeader } from './features/VesselSidebar/VesselSidebarHeader'
