+++ frontend/src/features/SideWindow/alerts_reportings/SilencedAlertsList.tsx
@@ -15,8 +15,8 @@ import { useMainAppDispatch } from '../../../hooks/useMainAppDispatch'
 import { useMainAppSelector } from '../../../hooks/useMainAppSelector'
 import { getDateTime } from '../../../utils'
 import SearchIconSVG from '../../icons/Loupe_dark.svg'
-import { Flag } from '../../vessel_list/tableCells'
-import { sortArrayByColumn, SortType } from '../../vessel_list/tableSort'
+import { Flag } from '../../VesselList/tableCells'
+import { sortArrayByColumn, SortType } from '../../VesselList/tableSort'
 
 import type { LEGACY_SilencedAlert } from '../../../domain/entities/alerts/types'
 import type { CSSProperties } from 'react'
