+++ frontend/src/features/VesselList/tableCells.tsx
