+++ frontend/src/features/favorite_vessels/FavoriteVessel.js
@@ -5,7 +5,7 @@ import { showVesselTrack } from '../../domain/use_cases/vessel/showVesselTrack'
 import { hideVesselTrack } from '../../domain/use_cases/vessel/hideVesselTrack'
 import { removeVesselFromFavorites } from '../../domain/shared_slices/FavoriteVessel'
 import { useDispatch } from 'react-redux'
-import { Flag } from '../vessel_list/tableCells'
+import { Flag } from '../VesselList/tableCells'
 import { ShowIcon } from '../commonStyles/icons/ShowIcon.style'
 import { HideIcon } from '../commonStyles/icons/HideIcon.style'
 import { CloseIcon } from '../commonStyles/icons/CloseIcon.style'
